.history/requirements_20250413160532.txt
# requirements.txt
# requirements.txt

pip install -r requirements.txt
asgiref
attrs
awsebcli
beautifulsoup4
boto3
botocore
cachetools
cassandra-driver
cement
certifi
channels
channels-redis
charset-normalizer
click
colorama
constantly
coreapi
coreschema
crispy-bootstrap4
crispy-bootstrap5
cryptography
daphne
decorator
defusedxml
diff-match-patch
dj-database-url

django-admin-sortable2
django-anymail
django-bootstrap4
django-ckeditor
django-ckeditor-5
django-cors-headers
django-crispy-forms
django-dbbackup
django-dotenv
django-environ
django-import-export
django-jazzmin

django-recaptcha==3.0.0
django-rest-auth
django-rest-framework
django-six
django-static-fontawesome
django-storages
django-user-agents
djangorestframework
djangorestframework-simplejwt
drf-yasg
environs
geoip2
geomet
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gunicorn
httplib2
humanize
hyperlink
idna
imageio
imageio-ffmpeg
Jinja2
Markdown
MarkupPy
MarkupSafe
marshmallow
maxminddb

msal

oauthlib
packaging

proglog
proto-plus
protobuf
psycopg2
pyasn1
pyasn1-modules
pycountry
PyJWT
python-dotenv
python3-openid
pytz
PyYAML
redis
requests
requests-oauthlib
rjsmin
rsa
s3transfer
service-identity
shortuuid
six
sqlparse
stripe
tqdm
twilio
tzdata
uritemplate
urllib3
whitenoise
django-extensions
razorpay